LPS-100795 Move SearchContext
