+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/management-toolbar/search/SearchBar.es.js
@@ -13,9 +13,9 @@
  */
 
 import React, {useContext} from 'react';
+import SearchContext from './SearchContext.es';
 import SearchInput from './SearchInput.es';
 import SearchSort from './SearchSort.es';
-import {SearchContext} from '../../table/SearchContext.es';
 
 export default ({columns, totalCount}) => {
 	const {
