+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/management-toolbar/search/SearchContext.es.js
@@ -62,4 +62,6 @@ const reducer = (state, action) => {
 	}
 };
 
-export {SearchContext, reducer};
+export default SearchContext;
+
+export {reducer};
