+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/management-toolbar/search/SearchSubnavigationBar.es.js
@@ -14,7 +14,7 @@
 
 import React, {useContext} from 'react';
 import Button from '../../button/Button.es';
-import {SearchContext} from '../../table/SearchContext.es';
+import SearchContext from './SearchContext.es';
 import lang from '../../../utils/lang.es';
 
 export default ({totalCount}) => {
