+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/components/table/TableWithPagination.es.js
@@ -14,10 +14,10 @@
 
 import React, {useContext} from 'react';
 import {ClayPaginationWithBar} from '@clayui/pagination';
-import {SearchContext} from './SearchContext.es';
 import {withEmpty} from './EmptyState.es';
 import Table from './Table.es';
 import {withLoading} from '../loading/Loading.es';
+import SearchContext from '../management-toolbar/search/SearchContext.es';
 
 const TableWithPagination = ({actions, columns, items, totalCount}) => {
 	const {
